Resolve NarrowCalculation and OperatorPrecedence from Error Prone (#6484)
* Resolve NarrowCalculation from Error Prone

This commit also changes to handle NarrowCalculation as errors.


* Resolve OperatorPrecedence from Error Prone

This commit also changes to handle OperatorPrecedence as errors.


---------

+++ micrometer-core/src/test/java/io/micrometer/core/instrument/composite/CompositeMeterRegistryTest.java
@@ -476,7 +476,7 @@ class CompositeMeterRegistryTest {
         assertThat(executor.awaitTermination(1L, TimeUnit.SECONDS)).isTrue();
         for (int i = 0; i < tagCount; i++) {
             assertThat(this.composite.find(meterName).tag(tagName, String.valueOf(i)).counter().count())
-                .isEqualTo(count / tagCount);
+                .isEqualTo((double) count / tagCount);
         }
     }
